全体編集

+++ db/migrate/20190912232419_add_reset_password_token_index_to_spree_users.solidus_auth.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
+
 # This migration comes from solidus_auth (originally 20190125170630)
 
+# AddResetPasswordTokenIndexToSpreeUsers
 class AddResetPasswordTokenIndexToSpreeUsers < SolidusSupport::Migration[4.2]
   # We're not using the standard Rails index name because somebody could have
   #  already added that index to the table. By using a custom name we ensure
@@ -28,8 +30,6 @@ class AddResetPasswordTokenIndexToSpreeUsers < SolidusSupport::Migration[4.2]
   end
 
   def down
-    if custom_index_exists?
-      remove_index :spree_users, name: custom_index_name
-    end
+    remove_index :spree_users, name: custom_index_name if custom_index_exists?
   end
 end
